+++ examples/locations/make_enum.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env/python
 
 """
-    make_enum.py -- make enumerations for adding degrees
+    make_enum.py -- make enumerations for locations
 """
 
 __author__ = "Michael Conlon"
@@ -42,58 +42,17 @@ def main():
     #   person
 
     query = """
-    SELECT ?short ?vivo
-    WHERE {
-      ?vivo a foaf:Person .
-      ?vivo rdfs:label ?short .
+    SELECT (MIN (?xlabel) AS ?short) ?vivo
+    WHERE
+    {
+          ?vivo rdf:type vivo:GeographicLocation .
+          ?vivo rdfs:label ?xlabel .
     }
+    GROUP BY ?vivo
     ORDER BY ?short
     """
 
-    create_enum("person_enum.txt", query, parms)
-
-    #   school
-
-    query = """
-    SELECT ?short ?vivo
-    WHERE {
-      ?vivo a vivo:University .
-      ?vivo rdfs:label ?short .
-    }
-    ORDER BY ?short
-    """
-
-    create_enum("school_enum.txt", query, parms)
-
-    #   date
-
-    query = """
-    SELECT ?short ?vivo
-    WHERE {
-      ?vivo a vivo:DateTimeValue .
-      ?vivo vivo:dateTimePrecision vivo:yearMonthPrecision .
-      ?vivo vivo:dateTime ?short .
-    }
-    ORDER BY ?short
-    """
-
-    create_enum("date_enum.txt", query, parms, trim=7)
-
-    #   academic degree
-
-    query = """
-    SELECT ?short ?vivo
-    WHERE {
-      ?vivo a vivo:AcademicDegree .
-      ?vivo rdfs:label ?short .
-    }
-    ORDER BY ?short
-    """
-
-    create_enum("degree_enum.txt", query, parms)
-
-    print datetime.now(), "End"
-    return None
+    create_enum("location_enum.txt", query, parms)
 
 
 if __name__ == "__main__":